% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/celebrate.r
\name{celebrate}
\alias{celebrate}
\title{Celebrating software birthdays}
\usage{
celebrate(github_username)
}
\arguments{
\item{github_username}{Character vector of one Github username handle (e.g., "some_username_here").}
}
\value{
Returns data frame with repository name, age, creation date, and time remaining until upcoming birthday.
}
\description{
Accesses all public repositories belonging to a specific Github user and prints each repository's respective name, age, creation date of initial commit (i.e., repository birthday date), and time remaining until upcoming birthday.
}
\examples{
celebrate("Bohdan-Khomtchouk")

}
