% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getData.R
\name{getData}
\alias{getData}
\title{getData}
\usage{
getData(
  id_df,
  tempdir = NA,
  veckeys = NA,
  telonic_usrs = NA,
  telonic_pass = NA,
  ATS_usrs = NA,
  ATS_pass = NA,
  lotek_usrs = NA,
  lotek_pass = NA,
  tzone = "America/Los_Angeles",
  subsetmonth = "02"
)
}
\arguments{
\item{id_df}{lookup table of animal IDs and serial numbers}

\item{tempdir}{temporary directory for storing data}

\item{veckeys}{path to vectronic keys}

\item{telonic_usrs}{vector of telonic usernames}

\item{telonic_pass}{vector of telonic passwords}

\item{ATS_usrs}{vector of ATS usernames}

\item{ATS_pass}{vector of ATS passwords}

\item{lotek_usrs}{vector of Lotek usernames}

\item{lotek_pass}{vector of Lotek passwords}

\item{tzone}{timezone of interest}

\item{subsetmonth}{month to subset GPS data}
}
\value{
Does not return any values. This function will save maps of the most recent locations in KML and GPS locations as well as the last three days of locations and movement in a leaflet map
}
\description{
pulls data for all target animals
}
