% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/happybirthday-package.R
\docType{package}
\name{happybirthday-package}
\alias{happybirthday}
\alias{happybirthday-package}
\title{happybirthday: Estimate movement metrics and create reports for bighorn sheep}
\description{
This package will download and clean gps data as well as create movement metrics and estimate parturition dates for mule deer, bighorn sheep, moose, and elk
}
\author{
\strong{Maintainer}: Katey Huggler \email{khuggler@uidaho.edu} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
