% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeTables.R
\name{makeTables}
\alias{makeTables}
\title{makeTables}
\usage{
makeTables(gpsdat, id_df, tempdir)
}
\arguments{
\item{gpsdat}{data.frame of gps data}

\item{id_df}{data.frame of lookup table}

\item{tempdir}{temporary directory to save files}
}
\value{
No return value. Function will save tables for rendering in rmarkdown
}
\description{
create tables of most recent 12 locations
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[sp]{coordinates}}, \code{\link[sp]{is.projected}}, \code{\link[sp]{spTransform}}
}
