% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moveMetrics.R
\name{moveMetrics}
\alias{moveMetrics}
\title{moveMetrics}
\usage{
moveMetrics(gpsdat, subsetmonth)
}
\arguments{
\item{gpsdat}{data.frame of gps data from the getData.R function}

\item{subsetmonth}{month to start subsetting GPS data. If GPS collars have recently been deployed, suggest using a month early in the year}
}
\value{
returns a data.frame with rolling statistics of movement metrics
}
\description{
calculate movement metrics including: movement rate, rolling home range size, residence time, and visitation rate
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[amt]{track}}, \code{\link[amt]{summarize_sampling_rate}}, \code{\link[amt]{reexports}}
\code{\link[adehabitatHR]{mcp}}
\code{\link[recurse]{getRecursions}}
\code{\link[zoo]{rollmean}}
}
