% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renderMarkdown.R
\name{renderMarkdown}
\alias{renderMarkdown}
\title{renderMarkdown}
\usage{
renderMarkdown(tempdir, markdowndir)
}
\arguments{
\item{tempdir}{temporary directory to store output}

\item{markdowndir}{directory to the Rmarkdown file for creating reports}
}
\value{
returns a pdf in tempdir with all collated plots and recent locations
}
\description{
render Rmarkdown file with plots and most recent locatiosn
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[rmarkdown]{render}}
}
