% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getData.R
\name{getData}
\alias{getData}
\title{getData}
\usage{
getData(id_df, tempdir, veckeys = NA)
}
\arguments{
\item{id_df}{data.frame that has the animal ids and serial numbers}

\item{tempdir}{character. path to temporary directory for files to be saved. If this path does not exist it will be created for you.}

\item{veckeys}{if Vectronic data is downloaded, path to Vectronic keys is needed, Default: NA}
}
\value{
Returns a data.frame with all gps data for a particular study area
}
\description{
This function downloads GPS data for a specified population and appends animal IDs
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[processx]{run}}
\code{\link[collar]{ats_login}}, \code{\link[collar]{get_paths}}, \code{\link[collar]{fetch_vectronics}}
\code{\link[dplyr]{bind}}
}
