% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeMaps.R
\name{makeMaps}
\alias{makeMaps}
\title{makeMaps}
\usage{
makeMaps(tempdir, gpsdat, id_df)
}
\arguments{
\item{tempdir}{temporary directory for storing maps}

\item{gpsdat}{data.frame of GPS data}

\item{id_df}{lookup table of animal IDs and serial numbers}
}
\value{
Does not return any values. This function will save maps of the most recent locations in KML and GPS locations as well as the last three days of locations and movement in a leaflet map
}
\description{
create maps of most recent locations and movement for bighorn sheep
}
\details{
some details here
}
\examples{
\dontrun{
if(interactive()){
 makeMaps(tempdir = 'directory', gpsdat = gpsdat, id_df = id_df)
 }
}
}
\seealso{
\code{\link[assertthat]{assert_that}}
\code{\link[sp]{coordinates}}, \code{\link[sp]{is.projected}}
\code{\link[purrr]{keep}}, \code{\link[purrr]{map}}
\code{\link[htmlwidgets]{saveWidget}}
}
