% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runUpdate.R
\name{runUpdate}
\alias{runUpdate}
\title{runUpdate}
\usage{
runUpdate(
  id_df,
  tempdir,
  veckeys = NA,
  telonic_usrs = NA,
  telonic_pass = NA,
  ATS_usrs = NA,
  ATS_pass = NA,
  lotek_usrs = NA,
  lotek_pass = NA,
  tzone = "America/Los_Angeles",
  subsetmonth = "02",
  study = NA,
  markdowndir = NA
)
}
\arguments{
\item{id_df}{data.frame of lookup information}

\item{tempdir}{temporary directory to save files}

\item{veckeys}{directory to vectronics key files}

\item{telonic_usrs}{vector of telonic usernames}

\item{telonic_pass}{vector of telonic passwords. The order of passwords must match that of the usernames}

\item{ATS_usrs}{vector of ATS usernames}

\item{ATS_pass}{vector of ATS passwords. The order of passwords must match that of the usernames}

\item{lotek_usrs}{vector of Lotek usernames}

\item{lotek_pass}{vector of Lotek passwords. The order of passwords must match that of the usernames}

\item{tzone}{time zone of vectronics data if applicable. Options are "America/Los_Angeles" or "America/Denver". Default is "America/Los_Angeles"}

\item{subsetmonth}{beginning month to subset GPS data to. Default is 2 (February)}

\item{study}{Name of study area. Only used for subject in update email}

\item{markdowndir}{directory to the Rmd file for the report}
}
\value{
final product is email with deliverables
}
\description{
wrapper function for running sheep update and emailing
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[happybirthday]{getData}}, \code{\link[happybirthday]{makeMaps}}, \code{\link[happybirthday]{moveMetrics}}, \code{\link[happybirthday]{makeMarkdown}}, \code{\link[happybirthday]{renderMarkdown}}
\code{\link[mailR]{send.mail}}
}
